requirements.txt
actionlib==1.14.0
angles==1.9.13
apriltag==0.0.16
asttokens==2.4.1
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
chardet==3.0.4
comm==0.2.2
constantly==15.1.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
debugpy==1.8.7
decorator==4.4.2
defusedxml==0.6.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
docutils==0.16
dynamic-reconfigure==1.7.3
empy==3.3.2
executing==2.1.0
filelock==3.0.12
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib_metadata==8.5.0
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyboard==0.13.5
kiwisolver==1.0.1
laser_geometry==1.6.7
lxml==4.5.0
lz4==3.0.2+dfsg
matplotlib==3.1.2
matplotlib-inline==0.1.7
message-filters==1.17.0
moveit-commander==1.1.15
moveit-core==1.1.15
moveit-ros-planning-interface==1.1.15
moveit-ros-visualization==1.1.15
mpi4py==3.0.3
nest-asyncio==1.6.0
netifaces==0.10.4
nose==1.3.7
numpy==1.24.4
packaging==24.1
pandas==2.0.3
paramiko==2.6.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==5.5.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyassimp==4.1.4
pycairo==1.16.2
pycryptodomex==3.6.1
pydot==1.4.1
Pygments==2.18.0
PyHamcrest==1.9.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQRCode==1.2.1
PyQt5==5.14.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-gnupg==0.4.5
python-pcl==0.3
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2024.2
PyYAML==5.3.1
pyzmq==26.2.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.7
roman==2.0.0
rosapi==0.11.17
rosbag==1.17.0
rosboost-cfg==1.15.8
rosbridge-library==0.11.17
rosbridge-server==0.11.17
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==0.9.1
rosdistro-modules==0.9.1
rosgraph==1.17.0
rosinstall==0.7.8
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.5.1
rospkg-modules==1.5.1
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.25
scipy==1.10.1
sensor-msgs==1.13.1
service-identity==18.1.0
sip==4.19.21
six==1.14.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
soupsieve==1.9.5
srdfdom==0.6.4
stack-data==0.6.3
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
topic-tools==1.17.0
tornado==6.4.1
traitlets==5.14.3
Twisted==18.9.0
txaio==2.10.0
typing_extensions==4.12.2
tzdata==2024.2
u-msgpack-python==2.1
urdfdom-py==0.4.6
vcstools==0.1.42
wcwidth==0.2.13
webencodings==0.5.1
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.18
zipp==3.20.2
zope.interface==4.7.1
